refactored menu styles
