+++ src/App.css
@@ -577,7 +577,6 @@ section,
   font-family: var(--heading-font);
 }
 
-
 /*--------------------------------------------------------------
 # About Section
 --------------------------------------------------------------*/
@@ -680,111 +679,6 @@ section,
 .why-us .card-item:hover h4 a,
 .why-us .card-item:hover p {
   color: var(--contrast-color);
-}
-
-/*--------------------------------------------------------------
-# Menu Section
---------------------------------------------------------------*/
-.menu {
-  background-color: #0f0f0f;
-}
-
-.menu .menu-filters {
-  padding: 0;
-  margin: 0 auto;
-  list-style: none;
-  text-align: center;
-  border-radius: 50px;
-}
-
-.menu .menu-filters li {
-  color: var(--default-color);
-  margin-bottom: 10px;
-  cursor: pointer;
-  display: inline-block;
-  padding: 8px 12px 10px 12px;
-  font-size: 16px;
-  font-weight: 500;
-  line-height: 1;
-  transition: all ease-in-out 0.3s;
-  font-family: var(--heading-font);
-}
-
-.menu .menu-filters li:hover,
-.menu .menu-filters li.filter-active {
-  color: var(--accent-color);
-  filter: invert(59%) sepia(69%) saturate(407%) hue-rotate(6deg)
-    brightness(120%) contrast(87%);
-}
-
-.menu .menu-filters li:last-child {
-  margin-right: 0;
-}
-
-.menu .menu-item {
-  margin-top: 50px;
-  position: relative;
-}
-
-.menu .menu-img {
-  width: 80px;
-  border-radius: 50%;
-  float: left;
-  border: 3px solid color-mix(in srgb, var(--default-color), transparent 90%);
-  margin-right: 20px;
-}
-
-.menu .menu-content {
-  margin-left: 0px;
-  overflow: hidden;
-  display: flex;
-  justify-content: space-between;
-  position: relative;
-}
-
-.menu .menu-content::after {
-  content: "......................................................................"
-    "...................................................................."
-    "....................................................................";
-
-  position: absolute;
-  left: 20px;
-  right: 0;
-  top: -4px;
-  z-index: 1;
-  color: color-mix(in srgb, var(--default-color), transparent 60%);
-  font-family: Arial, "Helvetica Neue", Helvetica, sans-serif;
-}
-
-.menu .menu-content a {
-  color: var(--heading-color);
-  background: var(--background-color);
-  padding-right: 10px;
-  position: relative;
-  z-index: 3;
-  font-weight: 700;
-  transition: 0.3s;
-}
-
-.menu .menu-content a:hover {
-  color: var(--accent-color);
-}
-
-.menu .menu-content span {
-  background: var(--background-color);
-  color: var(--accent-color);
-  position: relative;
-  z-index: 3;
-  padding: 0 10px;
-  font-weight: 600;
-}
-
-.menu .menu-ingredients {
-  margin-left: 0px;
-  font-style: italic;
-  font-size: 14px;
-  font-family: var(--nav-font);
-  color: color-mix(in srgb, var(--default-color), transparent 50%);
 }
 
 /*--------------------------------------------------------------
