+++ src/components/Hero/Hero.css
@@ -1,6 +1,3 @@
-/*--------------------------------------------------------------
-# Hero Section
---------------------------------------------------------------*/
 .hero {
   width: 100%;
   min-height: 85vh;
@@ -35,20 +32,29 @@
   z-index: 3;
 }
 
+.hero .content-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+}
+
 .hero h2 {
   margin: 0;
   font-size: 75px;
   font-weight: 700;
+  text-shadow: 0 2px 9px rgb(255 255 255 / 65%);
 }
 
 .hero h2 span {
   color: var(--accent-color);
 }
 
-.hero p {
+.hero p.hero-subtitle {
   color: color-mix(in srgb, var(--default-color), transparent 20%);
   margin: 10px 0 0 0;
   font-size: 24px;
+  text-shadow: 0 2px 9px rgb(255 255 255 / 65%);
 }
 
 .hero .cta-btn {
@@ -84,7 +90,7 @@
     font-size: 50px;
   }
 
-  .hero p {
+  .hero p.hero-subtitle {
     font-size: 18px;
   }
 }
