+++ src/components/Hero/Hero.jsx
@@ -1,5 +1,5 @@
 import background from "../../assets/img/bg3.jpg";
-import "./Hero.css";
+import "./hero.css";
 
 const Hero = () => {
   return (
@@ -7,19 +7,11 @@ const Hero = () => {
       <img src={background} alt="" />
       <div className="container">
         <div className="row">
-          <div
-            className="d-flex flex-column align-items-center align-items-lg-center"
-            style={{ textAlign: "center" }}
-          >
-            <h2
-              style={{ textShadow: "0 2px 9px rgb(255 255 255 / 65%)" }}
-              className="main-title"
-            >
+          <div className="content-wrapper">
+            <h2 className="main-title">
               Welcome to <span>SLS Shisha</span>
             </h2>
-            <p style={{ textShadow: "0 2px 9px rgb(255 255 255 / 65%)" }}>
-              Smoke the best shisha in Japan!
-            </p>
+            <p className="hero-subtitle">Smoke the best shisha in Japan!</p>
           </div>
         </div>
       </div>
